requirements.txt
pygame-ce
fastapi
uvicorn
requests
numpy
psycopg[binary]
python-dotenv